% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{C6Gates}
\alias{C6Gates}
\title{A gate set for S. cerevisiae (W303 derivatives) on BD Accuri C6 cytometers}
\usage{
C6Gates()
}
\value{
a set of polygonGate objects
contains "yeastGate", "dipdoubletGate", "dipsingletGate", "hapdoubletGate", and "hapsingletGate"
}
\description{
A gate set for S. cerevisiae (W303 derivatives) on BD Accuri C6 cytometers
}
\examples{
C6Gates()
}

