% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noSummaryAnalysis.R
\name{createAnnotation}
\alias{createAnnotation}
\title{Create an annotation dataframe}
\usage{
createAnnotation(yourFlowSet)
}
\arguments{
\item{yourFlowSet}{}
}
\value{
annotation_df
}
\description{
Create an annotation dataframe
}

