% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dipsingletGate}
\alias{dipsingletGate}
\title{A gate for the set of all diploid singlet yeast cells}
\format{
formal class polygonGate
}
\usage{
data(dipsingletGate)
}
\description{
Typically set in FSC.A by FSC.H space
Diploids are typically 5um x 6um ellipsoids while haploids are typically
4um x 4um spheroids. As a result, diploids are longer and you get a
larger 'area/volume'.
}
\keyword{datasets}
