% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{flowFrame.gettime}
\alias{flowFrame.gettime}
\title{Get the time at which at flowFrame began collection}
\usage{
flowFrame.gettime(flowframe)
}
\arguments{
\item{flowframe}{}
}
\description{
Get the time at which at flowFrame began collection
}

