% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\docType{data}
\name{gateEnv}
\alias{gateEnv}
\title{Create the gate environment}
\format{An object of class \code{environment} of length 5.}
\usage{
gateEnv
}
\value{
gate environment
}
\description{
This environment will allow the gates to be accessible throughout the package and prevent
interference with the global environment.
}
\examples{
loadGates()
ls(gateEnv)
}
\keyword{datasets}

