% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{get_time}
\alias{get_time}
\title{Get the time at which at flowFrame began collection}
\usage{
get_time(flowframe)
}
\arguments{
\item{flowframe}{The \code{flowFrame} for which you would like the initial time}
}
\value{
\code{numeric} time value in minutes
}
\description{
Get the time at which at flowFrame began collection
}
\examples{
plate1<-read.flowSet(path = system.file("extdata", "ss_example", package = "flowTime"),alter.names = TRUE)
get_time(plate1$A01.fcs)
}

