% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{getTime}
\alias{getTime}
\title{Get the time at which at flowFrame began collection}
\usage{
getTime(flowframe)
}
\arguments{
\item{flowframe}{The \code{flowFrame} for which you would like the initial
time}
}
\value{
\code{numeric} time value in minutes
}
\description{
Get the time at which at flowFrame began collection
}
\examples{
plate1<-read.flowSet(path = system.file("extdata", "ss_example", package =
"flowTime"),alter.names = TRUE)
getTime(plate1$A01.fcs)
}

