% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hapdoubletGate}
\alias{hapdoubletGate}
\title{A gate for the set of all haploid doublets}
\format{formal class polygonGate}
\usage{
data(hapdoubletGate)
}
\description{
A gate for the set of all haploid doublets
}
\keyword{datasets}

