% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\docType{data}
\name{hapsingletGate}
\alias{hapsingletGate}
\title{A gate for the set of all haploid singlets}
\format{formal class polygonGate}
\usage{
hapsingletGate
}
\description{
A gate for the set of all haploid singlets
}
\keyword{datasets}

