% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{listGates}
\alias{listGates}
\title{List existing gate sets saved in package memory}
\usage{
listGates()
}
\value{
lists the gates files saved within the flowTime package
}
\description{
List existing gate sets saved in package memory
}
\examples{
listGates()
}

