% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{loadGates}
\alias{loadGates}
\title{Load a yeast gate file}
\usage{
loadGates(gatesFile = "defaultGates.RData")
}
\arguments{
\item{gatesFile}{the gates file to be loaded into memory}
}
\value{
gate objects created in the current environment
}
\description{
Loads a set of yeast gates into active memory to be used in analysis functions
}
\examples{
loadGates()
}

