% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{meanMedianSD}
\alias{meanMedianSD}
\title{Summary statistic columns for a flow frame}
\usage{
meanMedianSD(frame)
}
\arguments{
\item{frame}{a \code{flowFrame}}
}
\value{
a matrix with a single row for the flow frame and mean, median, and
sd columns for each column of the expression measurements
}
\description{
Summary statistic columns for a flow frame
}
\examples{
plate1<-read.flowSet(path = system.file("extdata", "ss_example", package =
"flowTime"), alter.names = TRUE)
meanMedianSD(plate1@frames$A01.fcs)
fsApply(plate1, meanMedianSD)
}
