% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{qa.gating}
\alias{qa.gating}
\title{Quality assurance check
Check whether a flowSet (or a single flowFrame) contains empty values, in which case normalization may fail (divide by zero).
This is particularly useful for removing wash wells from a flowSet.}
\usage{
qa.gating(x, threshold = 100)
}
\arguments{
\item{x}{\code{flowSet} or \code{flowFrame} to be checked}

\item{threshold}{\code{flowFrames} with counts below this threshold will be identified. Defaults to \code{100}.}
}
\description{
Quality assurance check
Check whether a flowSet (or a single flowFrame) contains empty values, in which case normalization may fail (divide by zero).
This is particularly useful for removing wash wells from a flowSet.
}

