% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAnalysis.R
\name{renameflcols}
\alias{renameflcols}
\title{Rename flsummary data}
\usage{
renameflcols(x, channel = "FL1.A", transform = FALSE)
}
\arguments{
\item{x}{\code{data frame} with columns to be renamed}

\item{channel}{\code{character} that channel that was previously summarized}

\item{transform}{\code{character} how was the data transformed? FALSE if it was not}
}
\value{
the full \code{data frame} with renamed columns
}
\description{
Rename flsummary data
}
\examples{
plate1 <- read.flowSet(path = system.file("extdata", "ss_example/", package = "flowTime"), alter.names= TRUE)
flsummary(plate1, )
renameflcols(plate1, channel = "FL3.A", transform = FALSE)

}

