% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{saveGates}
\alias{saveGates}
\title{Save a yeast gate set}
\usage{
saveGates(yeastgate = yeastGate, dipsingletgate = dipsingletGate,
  dipdoubletgate = dipdoubletGate, hapsingletgate = hapsingletGate,
  hapdoubletgate = hapdoubletGate, fileName = "defaultGates.Rdata")
}
\arguments{
\item{yeastgate}{a gate object defining the population of yeast cells}

\item{dipsingletgate}{a gate object defining the population of diploid singlet cells}

\item{dipdoubletgate}{a gate object defining the population of diploid doublet cells}

\item{hapsingletgate}{a gate object defining the population of haploid singlet cells}

\item{hapdoubletgate}{a gate object defining the population of haploid doublet cells}

\item{fileName}{name of the .Rdata file you would like to save these gates within}
}
\value{
a .RData file in the "extdata" folder of the package containing the specified gates
}
\description{
Save a yeast gate set
}
\examples{
loadGates()
saveGates()
}

