% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{setGates}
\alias{setGates}
\title{Sets specified file to default gates for use in other functions}
\usage{
setGates(gatesFile)
}
\arguments{
\item{gatesFile}{}
}
\description{
Sets specified file to default gates for use in other functions
}

