% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{SORPGates}
\alias{SORPGates}
\title{A gate set for S. cerevisiae (W303 derivatives) on BD Accuri SORP cytometer
This is custom built for the Nemhauser Lab. This is BD Accuri SORP flow cytometer with a CSampler 96-well plate adapter using an excitation wavelength of 514 nm and an emission detection filter at 545/35 nm.}
\usage{
SORPGates()
}
\value{
a set of polygonGate objects
contains "yeastGate", "dipdoubletGate", "dipsingletGate", "hapdoubletGate",
and "hapsingletGate"
}
\description{
A gate set for S. cerevisiae (W303 derivatives) on BD Accuri SORP cytometer
This is custom built for the Nemhauser Lab. This is BD Accuri SORP flow cytometer with a CSampler 96-well plate adapter using an excitation wavelength of 514 nm and an emission detection filter at 545/35 nm.
}
\examples{
SORPGates()
yeastGate@boundaries # Show gate parameters
}

