% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noSummaryAnalysis.R
\name{steadyState}
\alias{steadyState}
\title{Find significant differences in steady state fluorescence}
\usage{
steadyState(flowset, gated = F, ploidy = "diploid", only = "singlets")
}
\arguments{
\item{gated}{is the data already gated?}

\item{ploidy}{= the gate to subset your flowsets based on the ploidy of you strains and your cytometer.}

\item{only=}{'yeast', 'singlets', or 'doublets'}
}
\description{
Compares several replicates of strain-treatment combinations for statistically significant differences in steady-state fluorescence
}

