% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAnalysis.R
\name{summary.cyt}
\alias{summary.cyt}
\title{summary.cyt:
Gates a sample to all yeast, then singlet, then doublets
Does the work of singletsummary.cyt,doubletsummary.cyt,yeastsummary.cyt
Also calculates singlet to doublet ratio
Returns a list of data frames, e.g. output$singlets, output$doublets, etc.}
\usage{
\method{summary}{cyt}(flowset, transform = F, channel = "FL1.A",
  gated = F, ploidy = F, moments = F, split = F, only = F)
}
\arguments{
\item{gated}{is the data already appropriately gated?}

\item{only}{}
}
\description{
summary.cyt:
Gates a sample to all yeast, then singlet, then doublets
Does the work of singletsummary.cyt,doubletsummary.cyt,yeastsummary.cyt
Also calculates singlet to doublet ratio
Returns a list of data frames, e.g. output$singlets, output$doublets, etc.
}

