% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/gateCreator.R
\docType{data}
\name{yeastGate}
\alias{yeastGate}
\title{A gate for the set of all yeast cells}
\format{formal class polygonGate}
\usage{
yeastGate

yeastGate
}
\value{
yeast gate
}
\description{
Typically set in FSC.A by SSC.A space to exclued any debris
    FSC.A  SSC.A
   400000  10000
    50000  15000
    40000  80000
  1250000 600000
  2000000 500000

Typically set in FSC.A by SSC.A space to exclued any debris
    FSC.A  SSC.A
   400000  10000
    50000  15000
    40000  80000
  1250000 600000
  2000000 500000
}
\examples{
loadGates()
show(yeastGate)
}
\keyword{datasets}

