% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAnalysis.R
\name{addbs}
\alias{addbs}
\title{Add background subtraction to a summary data frame}
\usage{
addbs(flowData, column = "FL3.Amean", baseline = "noYFP")
}
\arguments{
\item{flowData}{the summary data frame of to be background subtracted}

\item{column}{the column containing the fluorescent measurement to be background subtracted}

\item{baseline}{\code{character} the name of the strain representing background fluorescent values}
}
\value{
A summary data frame with an additional column "column_bs" containing the background subtracted fluorescent values
}
\description{
Subtracts the background fluorescence of a given control strain from the chosen column.
}
\examples{
dat<-read.flowSet(path=system.file("extdata", "tc_example", package = "flowTime"),alter.names = TRUE)
annotation <- read.csv(system.file("extdata", "tc_example.csv", package = "flowTime"))
annotation[which(annotation$treatment == 0), 'strain'] <- 'background'
adat <- annotateFlowSet(dat, annotation)
loadGates(gatesFile = 'C6Gates.RData')
dat_sum <- summarizeFlow(adat, ploidy = 'diploid', only = 'singlets',channel = 'FL1.A')
dat_sum <- addbs(dat_sum, column = "FL1.Amean", baseline = "background")

}

