% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\docType{data}
\name{dipdoubletGate}
\alias{dipdoubletGate}
\title{A gate for the set of all diploid doublets}
\format{formal class polygonGate}
\usage{
dipdoubletGate
}
\description{
A gate for the set of all diploid doublets
}
\keyword{datasets}

