% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAnalysis.R
\name{flsummary}
\alias{flsummary}
\title{Get summary statistics for fluorescence or other data channels of a flowSet}
\usage{
flsummary(flowset, channel = "FL3.A", moments = FALSE)
}
\arguments{
\item{flowset}{the \code{flowSet} to create summary statistics for}

\item{channel}{\code{character} the data channel to summarize}

\item{moments}{\code{boolean} if \code{TRUE} then split each frame into
early, middle, and late events}
}
\value{
A \code{data frame} containing summary statistics (mean, median,
SD) for the specified fluorescent channel and time moments of the flowSet.
}
\description{
Get summary statistics for fluorescence or other data channels of a flowSet
}
\examples{
plate1 <- read.flowSet(path = system.file("extdata",
"ss_example", package = "flowTime"), alter.names = TRUE)
flsummary(plate1)
}
