% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\docType{data}
\name{hapdoubletGate}
\alias{hapdoubletGate}
\title{A gate for the set of all haploid doublets}
\format{formal class polygonGate}
\usage{
hapdoubletGate
}
\value{
haploid doublet gate
}
\description{
A gate for the set of all haploid doublets
}
\examples{
loadGates()
hapdoubletGate
}
\keyword{datasets}

