% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{loadGates}
\alias{loadGates}
\title{Load a yeast gate file
into active memory}
\usage{
loadGates(gatesFile = "defaultGates.RData")
}
\arguments{
\item{gatesFile}{}
}
\description{
Load a yeast gate file
into active memory
}

