% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{loadGates}
\alias{loadGates}
\title{Load a yeast gate file}
\usage{
loadGates(gatesFile = NULL, path = NULL, envir = environment())
}
\arguments{
\item{gatesFile}{the gates file to be loaded into memory, or path to the
gates file}

\item{path}{The path to the gates file. If 'NULL' this will look through
lazy loaded data for the gatesFile}

\item{envir}{The environment in which to load the gates}
}
\value{
gate objects created in the current environment
}
\description{
Loads a set of yeast gates into active memory to be used in
analysis functions
}
\examples{
loadGates(system.file("extdata/SORPGates.RData", package = "flowTime"))
}
