% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{ploidy}
\alias{ploidy}
\title{Ploidy: Tries to guess the ploidy of a given flowframe using FSC.A/FSC.H ratio.
Diploids are typically 5um x 6um ellipsoids while haploids are typically 4um x 4um spheroids. As a result, diploids are longer and you get a larger 'area/volume' FSC.A. 'Width' might also be useful.}
\usage{
ploidy(flowframe)
}
\arguments{
\item{flowframe}{}
}
\description{
Ploidy: Tries to guess the ploidy of a given flowframe using FSC.A/FSC.H ratio.
Diploids are typically 5um x 6um ellipsoids while haploids are typically 4um x 4um spheroids. As a result, diploids are longer and you get a larger 'area/volume' FSC.A. 'Width' might also be useful.
}

