% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{polygate}
\alias{polygate}
\title{Create a polygon gate}
\usage{
polygate(x, y, filterID = "newGate", channels = c("FSC.A", "FSC.H"))
}
\arguments{
\item{channels}{}
}
\description{
Create a polygon gate
}

