% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryAnalysis.R
\name{renameflcols}
\alias{renameflcols}
\title{rename flsummary data
keeping it separate for ease of use
probably slow due to for loop}
\usage{
renameflcols(x, channel = "FL1.A", transform = F)
}
\arguments{
\item{transform}{}
}
\description{
rename flsummary data
keeping it separate for ease of use
probably slow due to for loop
}

