% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{saveGates}
\alias{saveGates}
\title{Save a yeast gate set}
\usage{
saveGates(yeastgate = yeastGate, dipsingletgate = dipsingletGate,
  dipdoubletgate = dipdoubletGate, hapsingletgate = hapsingletGate,
  hapdoubletgate = hapdoubletGate, fileName = "defaultGates.Rdata")
}
\arguments{
\item{file}{}
}
\description{
Save a yeast gate set
}

