% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gateCreator.R
\name{setGates}
\alias{setGates}
\title{Set default gates}
\usage{
setGates(gatesFile = "defaultGates.RData")
}
\arguments{
\item{gatesFile}{the full name of the gates file you would like to set as default (e.g. 'C6Gates.RData')}
}
\value{
overwrites the defaultGates.RData file with the specified gates file
}
\description{
Sets specified file to default gates for use in other functions
}
\examples{
loadGates("defaultGates.RData")
setGates("defaultGates.RData")
}

