% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeastGate}
\alias{yeastGate}
\title{A gate for the set of all yeast cells}
\format{formal class polygonGate}
\usage{
data(yeastGate)
}
\description{
Typically set in FSC.A by SSC.A space to exclued any debris
}
\keyword{datasets}

